fix: fixed cpath mapping file for collection attribute to use index 1 rather than index 0.
in the new cpath pull of the data, the AL Knowledge Portal value is
located at index 1 for the collection, not index 0 anymore. Index 0
contains values for if its part of the
progressive-supranuclear-palsy-collection which gives the name value of
Neurodegenrative disease, not ALS Knowledge Portal which is what we
want.